Refactor Phase 4d: extract field derivations (7 groups + facade)
Splits the ~30 _derive_* methods on ConnectorComparator into 7 grouped
classes under src/connect_migrate/mapper/properties/derivations/, plus a
DerivationGroup base and a FieldDeriver facade.

Layout:

  base.py                       DerivationGroup base class. Holds the
                                shared template helpers (_is_placeholder,
                                _extract_placeholder_name,
                                _resolve_template_default,
                                _get_template_default_value) so every
                                subclass inherits them. Each subclass
                                declares its DERIVATIONS dispatch table
                                (config-name -> method-name) and inherits
                                derive() / handles().
  connection_fields.py          ConnectionFieldDeriver (10 methods):
                                connection.url, host, port, user,
                                password, database, db.name,
                                db.connection.type, ssl.server.cert.dn,
                                database_server_name.
  format_fields.py              FormatFieldDeriver (6 methods): input/
                                output key/data/value formats.
  auth_fields.py                AuthFieldDeriver: authentication_method
                                + ssl.mode.
  redis_fields.py               RedisFieldDeriver: hostname, port,
                                ssl.mode.
  servicebus_fields.py          ServiceBusFieldDeriver: namespace +
                                three SAS-related keys.
  csfle_fields.py               CsfleFieldDeriver: csfle_enabled,
                                csfle_on_failure.
  schema_registry_fields.py     SchemaRegistryFieldDeriver:
                                subject-name-strategy variants.
  __init__.py                   FieldDeriver facade. Owns one instance
                                of each group; exposes lookup(name) ->
                                Optional[Callable] (matches the legacy
                                _derive_* signature, suitable as the
                                derivation_resolver callback for
                                ConfigDefProcessor) and a one-shot
                                derive() that dispatches and invokes.

The methods kept their bodies byte-identical apart from one rewrite:
self._jdbc_url_parser -> self.jdbc_url_parser (the new groups expose
the parser as a public attribute).

ConnectorComparator wires the facade in __init__ and now uses it from
both _generate_fm_config (top-level dispatch) and ConfigDefProcessor
(via the derivation_resolver callback). The class-internal dispatch
table _get_config_derivation_method and the four template helpers move
to the new home, alongside the 29 _derive_* methods that previously
filled most of the class body.

Deleted from ConnectorComparator (34 methods total):
  _get_config_derivation_method, _is_placeholder,
  _extract_placeholder_name, _resolve_template_default,
  _get_template_default_value, and all 29 _derive_*.

After this phase ConnectorComparator is down to 13 methods (from 131 in
the original file at the start of the refactor) and 1,517 lines (from
4,108). The class rename to ConnectorMapper and the _generate_fm_config
split into MappingPipeline are intentionally deferred.

No behavior change; --help unchanged; FieldDeriver dispatch verified
end-to-end against the original dispatch table entries.
